Update 012_while2.zig
 changed u32 to u16 (cuz we know the number won't change beyond u16 unless we broke things)

u16 = between 0 and 65,535.

the exercise have u32 (its not related to the current exercise )

+++ exercises/012_while2.zig
@@ -21,7 +21,7 @@
 const std = @import("std");
 
 pub fn main() void {
-    var n: u32 = 2;
+    var n: u16 = 2;
 
     // Please set the continue expression so that we get the desired
     // results in the print statement below.
